Bug Fixes
Bug Fixes and added Movable Interface

+++ cTimer.h
@@ -31,5 +31,5 @@ class cTimer
 		/**
 		*	@returns the DeltaTime.
 		*/
-		int getDelta();
+		double getDelta();
 };
